random_colors/requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
webcolors>=1.13.0